# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation VisonLand Affiliate",
        "description": "API này cho phép người dùng đọc thêm về các API được build trong dự án!",
        "contact": {
            "email": "pbao.business@gmail.com"
        },
        "version": "2"
    },
    "servers": [
        {
            "url": "http://localhost:8080/VisonLandAffiliate/"
        }
    ],
    "paths": {
...
        "/api/v1/address/ward": {
            "get": {
                "tags": [
                    "Ward"
                ],
                "summary": "Lấy danh sách phường xã",
                "description": "Lấy danh sách phường xã với phân trang.",
                "operationId": "6abaf3d1ffa21f5a22c3ab06177e77dd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại, page > 0.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số lượng bài viết trong 1 trang, limit > 0.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thực hiện thành công."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 7
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "parent 3"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "parent-3"
                                                    },
                                                    "children": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 8
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "child 3"
                                                                },
                                                                "slug": {
                                                                    "type": "string",
                                                                    "example": "child-3"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "has_prev": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "has_next": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Yêu cầu không hợp lệ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}